getDynLib-methods {inline}	R Documentation

Retrieve the dynamic library (or DLL) associated with a package of a function generated by cfunction

Description

The getDynLib function retrieves the dynamic library (or DLL) associated with a package or with a function generated by cfunction

Methods

signature(x = "CFunc")

Retrieves the dynamic library associated with the function generated by cfunction. The library is dynamically loaded if necessary.

signature(x = "CFuncList")

Retrieves the dynamic library associated with a set of functions generated by cfunction. The library is dynamically loaded if necessary.

signature(x = "character")

Retrieves the dynamic library of the given name. This typically refers to package names, but can be any name of the list returned by getLoadedDLLs

See Also

getLoadedDLLs, dyn.load

Examples

## Not run: 
getDynLib( "base" )

f <- cfunction( signature() , "return R_NilValue ;" )
getDynLib( f )

## End(Not run)

[Package inline version 0.3.15 Index]