package.skeleton-methods {inline}	R Documentation

Generate the skeleton of a package

Description

Generate the skeleton of a package

Methods

signature(name = "ANY", list = "ANY")

Standard method. See package.skeleton

signature(name = "character", list = "CFunc")

Method for a single generated by cfunction or cxxfunction

signature(name = "character", list = "CFuncList")

Method for a set functions generated by cfunction or cxxfunction

Examples

## Not run: 

fx <- cxxfunction( signature(x = "integer", y = "numeric" ) , '
	return ScalarReal( INTEGER(x)[0] * REAL(y)[0] ) ;
' )
package.skeleton( "foo", fx )

functions <- cxxfunction( 
	list( 
		ff = signature(), 
		gg = signature( x = "integer", y = "numeric" )
	), 
	c( "return R_NilValue ;", "return ScalarReal( INTEGER(x)[0] * REAL(y)[0] ) ;")
)
package.skeleton( "foobar", functions )

## End(Not run)

[Package inline version 0.3.15 Index]